# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  totalPaid: number;
  paymentCount: number;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  email?: string;
  notes?: string;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  email?: string;
  notes?: string;
}

export interface Payment {
  id: number;
  customerId: number;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  reason: string;
  paidAt: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface PaymentInput {
  /** @minimum 1 */
  customerId: number;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  reason: string;
  paidAt: string;
  notes?: string;
}

export type CustomerDetails = Customer & {
  payments: Payment[];
};

export interface RecentPayment {
  id: number;
  customerId: number;
  customerName: string;
  amount: number;
  reason: string;
  paidAt: string;
}

export interface DashboardSummary {
  totalRevenue: number;
  customerCount: number;
  paymentCount: number;
  recentPayments: RecentPayment[];
}

export type ListCustomersParams = {
search?: string;
};