# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  totalPaid: number;
  paymentCount: number;
  createdAt: Date;
}