# lib/api-zod/src/generated/types/customerDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Customer } from './customer';
import type { Payment } from './payment';

export type CustomerDetails = Customer & {
  payments: Payment[];
};