# lib/api-zod/src/generated/types/customerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  email?: string;
  notes?: string;
}