# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  phone?: string;
  email?: string;
  notes?: string;
}