# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecentPayment } from './recentPayment';

export interface DashboardSummary {
  totalRevenue: number;
  customerCount: number;
  paymentCount: number;
  recentPayments: RecentPayment[];
}