# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Payment {
  id: number;
  customerId: number;
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  reason: string;
  paidAt: Date;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}