# lib/api-zod/src/generated/types/recentPayment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecentPayment {
  id: number;
  customerId: number;
  customerName: string;
  amount: number;
  reason: string;
  paidAt: Date;
}